+++ pure_protobuf/__init__.py
@@ -1,1 +1,12 @@
+#!/usr/bin/env python3
+# coding: utf-8
+
+"""
+Python implementation of Protocol Buffers data types.
+
+`pure-protobuf` contributors © 2011-2019
+"""
+
+from __future__ import absolute_import
+
 from pure_protobuf.protobuf import *  # noqa: F401, F403